+++ src/user/test/user.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { UserController } from '../user.controller';
-import { UserService } from '../user.service';
+import { UserController } from '../kakao-user.controller';
+import { UserService } from '../kakao-user.service';
 
 describe('UserController', () => {
   let controller: UserController;
